Move instructions explorer to /agents; legacy agent pages to /old_agents
- /agents is now the Agents knowledge explorer (was /instructions); /instructions
  redirects there. /agents/new (creation wizard) stays under the new tree.
- Legacy data-source detail pages and list moved to /old_agents/* (data layout,
  internal links, and post-action redirects updated accordingly).
- KnowledgeExplorer supports deep-linking: /agents?agent=<id> opens that agent.
- Restore AgentSelector in the left nav as a nav-mode item: hover opens the
  agent dropdown, clicking the item goes to /agents, clicking an agent deep-links
  to /agents?agent=<id>. AgentFlyout / CommandPalette 'open agent' links updated.
- Align Playwright specs with the new routes (/agents, /old_agents).
