+++ frontend/layouts/data.vue
@@ -4,7 +4,7 @@
             <div class="w-full max-w-7xl px-4 ps-0 py-4">
 
                 <!-- Back link -->
-                <NuxtLink to="/agents" class="inline-flex items-center gap-1 text-xs text-gray-400 hover:text-gray-600 transition-colors">
+                <NuxtLink to="/old_agents" class="inline-flex items-center gap-1 text-xs text-gray-400 hover:text-gray-600 transition-colors">
                     <UIcon name="heroicons-chevron-left" class="w-3.5 h-3.5" />
                     All agents
                 </NuxtLink>
@@ -121,7 +121,7 @@
                         <p class="mt-1.5 text-sm text-gray-500 max-w-sm mx-auto">
                             This agent is private. Contact the owner or an admin to request access.
                         </p>
-                        <NuxtLink to="/agents" class="mt-4 inline-block text-sm text-blue-600 hover:underline">
+                        <NuxtLink to="/old_agents" class="mt-4 inline-block text-sm text-blue-600 hover:underline">
                             ← Back to agents
                         </NuxtLink>
                     </div>
@@ -134,7 +134,7 @@
                         <p class="mt-1.5 text-sm text-gray-500 max-w-sm mx-auto">
                             The agent you're looking for doesn't exist or has been removed.
                         </p>
-                        <NuxtLink to="/agents" class="mt-4 inline-block text-sm text-blue-600 hover:underline">
+                        <NuxtLink to="/old_agents" class="mt-4 inline-block text-sm text-blue-600 hover:underline">
                             ← Back to agents
                         </NuxtLink>
                     </div>
@@ -221,17 +221,17 @@ const tabs = computed(() =>
 )
 
 function tabTo(tabName: string) {
-    if (!id.value) return '/agents'
-    if (tabName === '') return `/agents/${id.value}`
-    return `/agents/${id.value}/${tabName}`
+    if (!id.value) return '/old_agents'
+    if (tabName === '') return `/old_agents/${id.value}`
+    return `/old_agents/${id.value}/${tabName}`
 }
 
 function isTabActive(tabName: string) {
     const path = route.path
     if (tabName === '') {
-        return path === `/agents/${id.value}` || path === `/agents/${id.value}/`
+        return path === `/old_agents/${id.value}` || path === `/old_agents/${id.value}/`
     }
-    return path === `/agents/${id.value}/${tabName}`
+    return path === `/old_agents/${id.value}/${tabName}`
 }
 
 const tableCount = computed(() =>
