+++ frontend/pages/instructions.vue
@@ -1,16 +1,14 @@
 <template>
-    <div class="text-sm h-full">
-        <div class="w-full h-full">
-            <KnowledgeExplorer />
-        </div>
-    </div>
+    <div />
 </template>
 
 <script setup lang="ts">
-import KnowledgeExplorer from '~/components/KnowledgeExplorer.vue'
-
+// The instructions page has moved to the Agents explorer at /agents.
+// Keep this route as a redirect so existing links/bookmarks still work.
 definePageMeta({
     auth: true,
     layout: 'default'
 })
+
+await navigateTo('/agents', { replace: true })
 </script>
