+++ frontend/pages/old_agents/[id]/monitoring.vue
