+++ frontend/pages/old_agents/[id]/tables.vue
