+++ frontend/pages/old_agents/[id]/tools.vue
