+++ frontend/tests/visibility/shared-access.spec.ts
@@ -19,7 +19,7 @@ test.describe('Shared page visibility (both admin and member)', () => {
   });
 
   test('admin can access instructions page', async ({ adminPage }) => {
-    await adminPage.goto('/instructions');
+    await adminPage.goto('/agents');
     await adminPage.waitForLoadState('domcontentloaded');
 
     await expect(adminPage.getByRole('heading', { name: 'Agents', exact: true }))
@@ -27,7 +27,7 @@ test.describe('Shared page visibility (both admin and member)', () => {
   });
 
   test('member can access instructions page', async ({ memberPage }) => {
-    await memberPage.goto('/instructions');
+    await memberPage.goto('/agents');
     await memberPage.waitForLoadState('domcontentloaded');
 
     await expect(memberPage.getByRole('heading', { name: 'Agents', exact: true }))
